+++ tests/dne.js
@@ -1,14 +1,12 @@
 import { RequestHandlerTest } from '../RequestHandlerTest.js';
-import { JS } from '@shgysk8zer0/consts/mimes.js';
+import { NOT_FOUND } from '@shgysk8zer0/consts/status.js';
 import { TestRequest } from '../TestRequest.js';
 
 const { error } = await RequestHandlerTest.runTests(
 	new RequestHandlerTest(
-		new TestRequest('http://example.com/api/script', {
-			headers: { 'Accept': JS }
-		}),
-		[RequestHandlerTest.shouldBeOk, RequestHandlerTest.shouldHaveContentType(JS)]
-	)
+		new TestRequest(new URL('./dne', 'http://localhost:8888')),
+		[RequestHandlerTest.shouldHaveStatus(NOT_FOUND)]
+	),
 );
 
 if (error instanceof Error) {