+++ tests/redirect.js
@@ -3,7 +3,7 @@ import { RequestHandlerTest } from '../RequestHandlerTest.js';
 const { error } = await RequestHandlerTest.runTests(
 	new RequestHandlerTest(
 		new Request('https://example.com/api/redirect', { redirect: 'manual' }),
-		RequestHandlerTest.shouldRedirect
+		RequestHandlerTest.shouldRedirectTo(new URLPattern({ pathname: '/api/echo' }))
 	)
 );